\name{print.summary.bayesplfm}
\alias{print.summary.bayesplfm}
\title{Printing summaries of Bayesian probabilistic latent feature analysis
}
\description{
Printing method for summaries of objects generated by \code{\link{summary.bayesplfm}}
}
\usage{
\method{print}{summary.bayesplfm}(x,\dots)
}
\arguments{
  \item{x}{Summary of Bayesian probabilistic latent feature analysis returned 
           by \code{\link{summary.bayesplfm}}}
  \item{\dots}{Further arguments are ignored}
}
\seealso{
\code{\link{summary.bayesplfm}}, \code{\link{bayesplfm}}
}
