\name{print.summary.plfm}
\alias{print.summary.plfm}
\title{Printing summaries of probabilistic latent feature analysis objects}
\description{Printing method for summaries of probabilistic latent feature analysis objects.}
\usage{\method{print}{summary.plfm}(x,\dots)}
\arguments{,
  \item{x}{Summary of a probabilistic latent feature analysis object returned by \code{\link{summary.plfm}}}
  \item{\dots}{Further arguments are ignored}
}
\seealso{
\code{\link{plfm}}, \code{\link{summary.plfm}}
}
