\name{print.summary.stepLCplfm}
\alias{print.summary.stepLCplfm}
\title{Printing summaries of \code{\link{stepLCplfm}} objects}
\description{Printing method for summaries of \code{\link{stepLCplfm}} objects.}
\usage{\method{print}{summary.stepLCplfm}(x,digits=2,\dots)}
\arguments{
  \item{x}{Summary of a list of latent class probabilistic latent feature analysis objects returned by \code{\link{summary.stepLCplfm}}.}
   \item{digits}{By default 2 significant digits are used for printing.}
   \item{\dots}{Further arguments are ignored.}
}
\seealso{
\code{\link{stepLCplfm}}, \code{\link{summary.stepLCplfm}}
}
