\name{print.summary.stepplfm}
\alias{print.summary.stepplfm}
\title{Printing summaries of \code{\link{stepplfm}} objects}
\description{Printing method for summaries of \code{\link{stepplfm}} objects.}
\usage{\method{print}{summary.stepplfm}(x,digits=2,\dots)}
\arguments{
  \item{x}{Summary of a list of probabilistic latent feature analysis object returned by \code{\link{summary.stepplfm}}.}
   \item{digits}{By default 2 significant digits are used for printing.}
   \item{\dots}{Further arguments are ignored.}
}
\seealso{
\code{\link{stepplfm}}, \code{\link{summary.stepplfm}}
}
